Add multi-task training modes
